+++ kerem-net-nest/src/Data/posts.json
@@ -1,17 +1,17 @@
 [
     {
         "id": "sad",
-        "userId": "",
-        "text": "",
+        "userId": "dfgsdfg",
+        "text": "sdfgsdfgsdfg",
         "comments": [
             {
-                "commentContent": "",
-                "userName": "",
-                "date": ""
+                "commentContent": "dsfgsdfg",
+                "userName": "sfdgsdfg",
+                "date": "sdfgsfg"
             }
         ],
         "likes": [],
-        "creatorName": "",
-        "date": ""
+        "creatorName": "sdfgsdfg",
+        "date": "sdfgsdf"
     }
 ]
