+++ kerem-net-nest/src/Interfaces/CreatePostDto.ts
@@ -1,0 +1,6 @@
+/* eslint-disable prettier/prettier */
+export class CreatePostDto {
+  userId: string;
+  text: string;
+  creatorName: string;
+}
