+++ kerem-net-nest/src/Interfaces/CreateUserDto.ts
@@ -1,0 +1,5 @@
+/* eslint-disable prettier/prettier */
+export class CreateUserDto {
+  username: string;
+  biography: string;
+}
