+++ kerem-net-nest/src/Interfaces/FollowUserDto.ts
@@ -1,0 +1,5 @@
+/* eslint-disable prettier/prettier */
+export class FollowUserDto {
+  followerId: string;
+  userId: string;
+}
