+++ kerem-net-nest/src/Interfaces/LikePostDto.ts
@@ -1,0 +1,5 @@
+/* eslint-disable prettier/prettier */
+export class LikePostDto {
+  userId: string;
+  postId: string;
+}
